+++ server/src/test/java/com/inventage/portal/gateway/proxy/middleware/VertxAssertions.java
@@ -2,6 +2,7 @@ package com.inventage.portal.gateway.proxy.middleware;
 
 import io.vertx.junit5.VertxTestContext;
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.function.Executable;
 
 public class VertxAssertions {
 
@@ -74,6 +75,18 @@ public class VertxAssertions {
     public static void fail(VertxTestContext testCtx, Throwable cause) {
         testCtx.verify(() -> {
             Assertions.fail(cause);
+        });
+    }
+
+    public static void assertDoesNotThrow(VertxTestContext testCtx, Executable executable, String message) {
+        testCtx.verify(() -> {
+            Assertions.assertDoesNotThrow(executable, message);
+        });
+    }
+
+    public static <T extends Throwable> void assertThrows(VertxTestContext testCtx, Class<T> expectedType, Executable executable, String message) {
+        testCtx.verify(() -> {
+            Assertions.assertThrows(expectedType, executable, message);
         });
     }
 }